// src/shared/api/models/getApiBundleOffersSoldIdTypeParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiBundleOffersSoldIdTypeParams = {
    /**
     * Start date in YYYY-MM-DD format
     */
    fromDate?: string;
    /**
     * End date in YYYY-MM-DD format
     */
    toDate?: string;
};